requirements.txt
Flask==2.3.0
Werkzeug==2.3.7
mysql-connector-python==8.0.33
pytest==7.3.1
selenium==4.9.0
pylint==2.17.4
